% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_df.R
\name{get_df}
\alias{get_df}
\title{get_df}
\usage{
get_df(infiles)
}
\arguments{
\item{infiles}{vector of filenames}
}
\description{
Returns a dataframe built from several sqlite files
}
\examples{
infile_1 <- system.file("extdata", "example.db", package = "pubcount")
infile_2 <- system.file("extdata", "dengue_paris.db", package = "pubcount")
get_df(c(infile_1, infile_2))
}
